% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_model_CREP.R
\name{annual_model_CREP}
\alias{annual_model_CREP}
\title{This function produces annual amounts of water use and profits.}
\usage{
annual_model_CREP(
  soil_weather_file = "./input_files/Well_SoilType_WeatherStation.csv",
  well_capacity_files = "./Well Capacity",
  econ_output_folder = "./Econ_output/results_with_CREP/annual_results",
  well_capacity_file_year = "./KS_DSSAT_output.csv",
  look_up_table = "lookup_table_all_years_2.rds",
  CREP_wells_data = "./CREP_wells.csv",
  first_year_of_simulation = 2000,
  default_well_capacity_col_name = "Well_Capacity(gpm)",
  missing_soil_types = "KS00000007",
  minimum_well_capacity = 0,
  maximum_well_capacity = 800,
  well_capacity_intervals = 10,
  first_year_of_GW = 1997,
  last_year_of_GW = 2007,
  irrigation_season_days = 70
)
}
\arguments{
\item{soil_weather_file}{is the file that contains the soil types for each well in the region. Defaults to "C:/Users/manirad/Dropbox/DSSAT subregions work pre-2018 annual meeting/subregion KS files/outputs_for_econ/Well_Soil Type.csv".}

\item{well_capacity_files}{is the directory where well capacity files are located. Defaults to "C:/Users/manirad/Downloads/test/Well Capacity".}

\item{econ_output_folder}{is the name of the output folder that contains irrigated acres, irrigation, and profits for each well. Defaults to "C:/Users/manirad/Downloads/test/Econ_output/KS_DSSAT_output.csv",}

\item{well_capacity_file_year}{is the name of the output file that contains irrigation for each well which will be used by MODFLOW. Defaults to "C:/Users/manirad/Downloads/test/KS_DSSAT_output.csv",}

\item{look_up_table}{is the lookup table for wells that are inside  the selected policy area.}

\item{CREP_wells_data}{is the file that contains CREP well ID's. This can be set to the wells that you want retired.}

\item{first_year_of_simulation}{is the first year that the hydro-economic simulation starts. Defaults to 2000.}

\item{default_well_capacity_col_name}{is the name of the well capacity column generated from the MODFLOW model. Defaults to 'Well_Capacity(gpm)' as this is the original column name we started with.}

\item{missing_soil_types}{is the soil type that is assigned to missing soil types for wells. Defaults to "KS00000007".}

\item{minimum_well_capacity}{is the minimum well capacity in the model. If well capacity falls below this capacity, it is set to this minimum. Defaults to 100  gallons per minute.}

\item{maximum_well_capacity}{is the maximum well capacity in the model. If well capacity falls above this capacity, it is set to this maximum. Defaults to 3000 gallons per minute.}

\item{first_year_of_GW}{is the first year that the GW model exists. This may be different than the first year of simulation. Defaults to 1997.}

\item{last_year_of_GW}{is the last  year that the GW model exists. This may be different than the last  year of simulation. Defaults to 2007.}

\item{irrigation_season_days}{Number of days in an irrigation season. Defaults to 70.}
}
\value{
returns the output table.
}
\description{
This function produces annual amounts of water use and profits.
}
