% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_optim.R
\name{FN_optim}
\alias{FN_optim}
\title{This function returns maximum expected profits for each quarter circle.}
\usage{
FN_optim(jj)
}
\arguments{
\item{jj}{is the loop that goes through the wells in the study area}
}
\value{
returns the output table.
}
\description{
This function returns maximum expected profits for each quarter circle.
}
