% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_optim2.R
\name{FN_optim2}
\alias{FN_optim2}
\title{This function maximizes the expected profit.}
\usage{
FN_optim2(jj = 1)
}
\arguments{
\item{jj}{is the loop number. Defaults to 1.}
}
\value{
returns the optimized values
}
\description{
This function maximizes the expected profit.
}
