% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_optim3.R
\name{FN_optim3}
\alias{FN_optim3}
\title{This function maximizes the expected profit.}
\usage{
FN_optim3(jj)
}
\arguments{
\item{jj}{is the loop number. Defaults to 1.}
}
\value{
returns the optimized values
}
\description{
This function maximizes the expected profit.
}
