% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_CREP_wells.R
\name{gen_CREP_wells}
\alias{gen_CREP_wells}
\title{This function generates the lookup table with a tax policy for KS.}
\usage{
gen_CREP_wells(
  seed = 1,
  number_of_CREP_wells = 12,
  well_id_file = "all_well_IDs.csv",
  output_file = "CREP_wells.csv"
)
}
\arguments{
\item{seed}{Determines the random draw. This value is updated within a loop for bootstrapping.}

\item{number_of_CREP_wells}{is the number of wells that will be retired as a result of the conservation policy. Defaults to 169.}

\item{well_id_file}{is the file that contains the well ID's.}

\item{output_file}{is the output file that includes well ID for retired wells.}
}
\value{
returns the output table.
}
\description{
This function generates the lookup table with a tax policy for KS.
}
