% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lookup_crop_insurance.R
\name{gen_lookup_crop_insurance}
\alias{gen_lookup_crop_insurance}
\title{This function generates outputs for different crop insurance parameters.}
\usage{
gen_lookup_crop_insurance(
  DSSAT_files = "./input_files/DSSAT_files",
  well_soil_file = "./input_files/Well_Soil Type.csv",
  well_capacity_files = "./Well Capacity",
  price_file = "./input_files/crop_prices.csv",
  fixed_cost_file = "./input_files/fixed_cost_input.csv",
  insparamfile = "./input_files/insparam.csv",
  parcel_APH_file = "./parcel_APH.csv",
  county_APH_file = "./county_ref.csv",
  cover_file = "./input_files/coverage_subsidy.csv",
  econ_output_folder = "./Econ_output/results_with_insurance/",
  hydrology_file_year = "./KS_DSSAT_output.csv",
  default_well_capacity_col_name = "Well_Capacity(gpm)",
  missing_soil_types = "KS00000007",
  pumping_cost = 3,
  soil_moisture_targets = c(25, 35, 45, 55, 65, 75),
  IFREQ_seq = 2,
  IFREQ_interpolate = 0.1,
  minimum_well_capacity = 0,
  maximum_well_capacity = 1000,
  first_year_of_GW = 1997,
  last_year_of_GW = 2008,
  irrigation_season_days = 70,
  first_year_of_simulation = 1999,
  insurance_subsidy_increase = 0,
  no_policy_indicator = 1,
  num_clusters = parallel::detectCores() - 2
)
}
\arguments{
\item{DSSAT_files}{is the directory where DSSAT files are located. Defaults to "C:/Users/manirad/Dropbox/DSSAT subregions work pre-2018 annual meeting/subregion KS files/outputs_for_econ/revised".}

\item{well_soil_file}{is the file that contains base soil file. Defaults to "C:/Users/manirad/Downloads/test/Well_Soil Type_generator_07.csv".}

\item{well_capacity_files}{is the file that contains base well capacity. Defaults to "C:/Users/manirad/Downloads/test/Well_Capacity_ganarator.csv".}

\item{price_file}{is the file that includes crop prices. Defaults to "C:/Users/manirad/Dropbox/DSSAT subregions work pre-2018 annual meeting/subregion KS files/crop_prices.csv".}

\item{fixed_cost_file}{is the file tha includes fixed costs (per acre) costs of production. Defaults to "C:/Users/manirad/Downloads/test/fixed_cost_input.csv".}

\item{insparamfile}{is the value of insurance parameters that are used to estimate the premium rate.}

\item{parcel_APH_file}{is the file that keeps track of the APH at the parcel-level.}

\item{county_APH_file}{is the file that keeps tracl of the APH at the county-level.}

\item{cover_file}{is the file that includes different coverage levels and their respective subsidies.}

\item{econ_output_folder}{is the folder where outputs will be written in.}

\item{hydrology_file_year}{is the file that communicates daily groundwater use for each well. This file is read by the MODFLOW model.}

\item{default_well_capacity_col_name}{is the name of the well capacity column generated from the MODFLOW model. Defaults to 'Well_Capacity(gpm)' as this is the original column name we started with.}

\item{missing_soil_types}{is the name of the soil type that will be assigned to wells that end up with an NA soil type. This is to avoid dropping wells from the analysis.}

\item{pumping_cost}{is the cost of pumping an acre-inch of groundwater excluding taxes. Defaults to 3.56 which is the cost of pumpin an acre-inch of groundwater in parts of Kansas.}

\item{soil_moisture_targets}{is the vector of soil moisture targets that are generated by the DSSAT model. (25, 35, 45, 55, 65, 75).}

\item{IFREQ_seq}{is the difference between two consequtive irrigation frequencies (IFREQ) in DSSAT. Defaults to 2.}

\item{IFREQ_interpolate}{is the size of interpolation interval. Defaults to 0.1 so that IFREQ_seq of 2 adds 0, 0.1, 0.2, ..., 1.9.}

\item{minimum_well_capacity}{is the minimum well capacity in the model. If well capacity falls below this capacity, it is set to this minimum. Defaults to 100  gallons per minute.}

\item{maximum_well_capacity}{is the maximum well capacity in the model. If well capacity falls above this capacity, it is set to this maximum. Defaults to 3000 gallons per minute.}

\item{first_year_of_GW}{is the first year that the GW model exists. This may be different than the first year of simulation. Defaults to 1997.}

\item{last_year_of_GW}{is the last  year that the GW model exists. This may be different than the last  year of simulation. Defaults to 2007.}

\item{irrigation_season_days}{Number of days in an irrigation season. Defaults to 70.}

\item{first_year_of_simulation}{is the first year that the hydro-economic simulation starts. Defaults to 2000.}

\item{insurance_subsidy_increase}{is the change in insurance subsidy. This can be positive or negative.}

\item{no_policy_indicator}{is an indicator that is equal to 0 if the policy is no crop insurance policy. Otherwise it's equal to 1.}

\item{num_clusters}{is the number of cores for parallelization.}
}
\value{
returns the output table.
}
\description{
This function generates outputs for different crop insurance parameters.
}
\examples{
\dontrun{
gen_lookup_tax(subsidy_amount = 2)
}
}
