% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lookup_tax_KS.R
\name{gen_lookup_tax_KS}
\alias{gen_lookup_tax_KS}
\title{This function generates the lookup table with a tax policy for KS.}
\usage{
gen_lookup_tax_KS(
  tax_amount = 0.1,
  DSSAT_files = "./input_files/DSSAT_files",
  soil_file = "./input_files/Well_Soil Type_generator_07.csv",
  well_capacity_file = "./input_files/Well_Capacity_ganarator.csv",
  price_file = "./input_files/crop_prices.csv",
  fixed_cost_file = "./input_files/fixed_cost_input.csv",
  pumping_cost = 3.56,
  default_well_capacity_col_name = "Well_Capacity(gpm)",
  soil_moisture_targets = c(25, 35, 45, 55, 65, 75),
  IFREQ_seq = 2,
  IFREQ_interpolate = 0.1
)
}
\arguments{
\item{tax_amount}{is the amount of tax on the unit of groundwater extracted. Defaults to 1.1.}

\item{DSSAT_files}{is the directory where DSSAT files are located. Defaults to "C:/Users/manirad/Dropbox/DSSAT subregions work pre-2018 annual meeting/subregion KS files/outputs_for_econ/revised".}

\item{soil_file}{is the file that contains base soil file. Defaults to "C:/Users/manirad/Downloads/test/Well_Soil Type_generator_07.csv".}

\item{well_capacity_file}{is the file that contains base well capacity. Defaults to "C:/Users/manirad/Downloads/test/Well_Capacity_ganarator.csv".}

\item{price_file}{is the file that includes crop prices. Defaults to "C:/Users/manirad/Dropbox/DSSAT subregions work pre-2018 annual meeting/subregion KS files/crop_prices.csv".}

\item{fixed_cost_file}{is the file tha includes fixed costs (per acre) costs of production. Defaults to "C:/Users/manirad/Downloads/test/fixed_cost_input.csv".}

\item{pumping_cost}{is the cost of pumping an acre-inch of groundwater excluding taxes. Defaults to 3.56 which is the cost of pumpin an acre-inch of groundwater in parts of Kansas.}

\item{default_well_capacity_col_name}{is the name of the well capacity column generated from the MODFLOW model. Defaults to 'Well_Capacity(gpm)' as this is the original column name we started with.}

\item{soil_moisture_targets}{is the vector of soil moisture targets that are generated by the DSSAT model. (25, 35, 45, 55, 65, 75).}

\item{IFREQ_seq}{is the difference between two consequtive irrigation frequencies (IFREQ) in DSSAT. Defaults to 2.}

\item{IFREQ_interpolate}{is the size of interpolation interval. Defaults to 0.1 so that IFREQ_seq of 2 adds 0, 0.1, 0.2, ..., 1.9.}
}
\value{
returns the output table.
}
\description{
This function generates the lookup table with a tax policy for KS.
}
\examples{
\dontrun{
gen_lookup_tax(tax_amount = 2)
}
}
