% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_output_subsidy_nolookup.R
\name{gen_output_subsidy_nolookup}
\alias{gen_output_subsidy_nolookup}
\title{This function produces annual well capacity for each well in each year.}
\usage{
gen_output_subsidy_nolookup(
  subsidy_amount = 2,
  subsidy_threshold = 200,
  well_capacity_files = "./Well Capacity",
  well_capacity_for_econ = "./Econ_output/well_capacity.csv",
  first_year_of_simulation = 2000,
  default_well_capacity_col_name = "Well_Capacity(gpm)"
)
}
\arguments{
\item{subsidy_amount}{is the amount of subsidy per acre-inch of groundwater extracted. Defaults to 1.}

\item{subsidy_threshold}{is the threshold of subsidy, i.e., the amount of groundwater extraction above which subsidy is zero. Defaults to 400.}

\item{well_capacity_for_econ}{is the file name that all the output well capacities are written into. Defaults to "C:/Users/manirad/Downloads/test/Econ_output/well_capacity.csv".}

\item{first_year_of_simulation}{is the first year that the hydro-economic simulation starts. Defaults to 2000.}

\item{default_well_capacity_col_name}{is the name of the well capacity column generated from the MODFLOW model. Defaults to 'Well_Capacity(gpm)' as this is the original column name we started with.}
}
\value{
returns the output table.
}
\description{
This function produces annual well capacity for each well in each year.
}
\examples{
\dontrun{
gen_lookup_subsidy(subsidy_amount = 2)
}
}
